Code refactor and add [Todo]
