+++ config/env.js
@@ -7,6 +7,7 @@ dotenv.config({
   path: pathName,
 });
 
+// 1. add default
 module.exports = {
   NODE_ENV: process.env.NODE_ENV || "development",
   MONGO_CONNECTION_URL: process.env.MONGO_CONNECTION_URL,
