+++ routes/userRoutes.js
@@ -1,6 +1,6 @@
 const router = require("express").Router();
 
-const { fileUploader } = require("../controller/userController");
+const { fileUploader } = require("../controller/fileController");
 const upload = require("../middleware/fileUploader");
 
 router.post("/", upload, fileUploader);
